Move quaternary expression to spark shim
